requirements.txt

# requirements.txt content:

requests>=2.28.0
mysql-connector-python>=8.0.32
python-dotenv>=1.0.0
schedule>=1.2.0
numpy>=1.21.0
pandas>=1.3.0
